Version 21.11.18: Reworked the way sprites get gathered, rescaled and rendered. Massive memory usage reduction (50% on test project). Code cleanup.
