+++ src/engine/rendering/SoftwareRenderer.java
@@ -1,5 +1,6 @@
 package engine.rendering;
 
+import engine.SpriteFactory;
 import engine.Vector2;
 import engine.gameobject.component.Camera;
 import engine.gameobject.component.Transform;
@@ -37,7 +38,7 @@ public class SoftwareRenderer {
     private ArrayList<SpriteRenderer> renderQueue;
     private float aspectRatio;
     private Camera camera;
-    private float verticalSpriteSizeTarget;
+    private ArrayList<Camera> cameras;
 
     /** Constructs a new SoftwareRenderer with the desired width and height of the rendering zone.
      *
@@ -64,6 +65,7 @@ public class SoftwareRenderer {
         window.getContentPane().addMouseListener(Input.getInstance().getMouseInput());
         sprites = new ArrayList<>();
         renderQueue = new ArrayList<>();
+        cameras = new ArrayList<>();
     }
 
     /** Cull, sort and render sprites to the rendering zone. Can clear the buffer with a determined color before rendering.
@@ -71,34 +73,31 @@ public class SoftwareRenderer {
      */
     public void update() {
 
-        if(camera == null) {
+        cull();
+        zSort();
 
-            Arrays.fill(rawFrameBuffer, CLEARCOLOR);
-        } else {
+        Arrays.fill(rawFrameBuffer, CLEARCOLOR);
 
-            cull();
-            zSort();
+        int x;
+        int y;
+        SpriteRenderer sprite;
+        Vector2 projectedPosition;
 
-            Arrays.fill(rawFrameBuffer, CLEARCOLOR);
+        for (int i = renderQueue.size() - 1; i > -1; i--) {
 
-            int x;
-            int y;
-            SpriteRenderer sprite;
-            Vector2 projectedPosition;
+            sprite = renderQueue.get(i);
+            projectedPosition = camera.worldToCamera(sprite.getGameObject().getTransform().position());
 
-            for (int i = renderQueue.size() - 1; i > -1; i--) {
+            BufferedImage spriteToRender = SpriteFactory.getInstance().getScaledSprite(sprite.getName(), sprite.getGameObject());
 
-                sprite = renderQueue.get(i);
-                projectedPosition = camera.worldToCamera(sprite.getGameObject().getTransform().position());
+            x = (int) ((width * projectedPosition.getX() * (camera.getMaxRenderArea().getX() - camera.getMinRenderArea().getX())) - spriteToRender.getWidth() / 2f);
+            x += (int)((float) width * camera.getMinRenderArea().getX());
 
-                x = (int) ((width * projectedPosition.getX() * (camera.getMaxRenderArea().getX() - camera.getMinRenderArea().getX())) - sprite.getScaledSprite().getWidth() / 2f);
-                x += (int)((float) width * camera.getMinRenderArea().getX());
+            y = (int) ((height * projectedPosition.getY() * (camera.getMaxRenderArea().getY() - camera.getMinRenderArea().getY()) - spriteToRender.getHeight() / 2f));
+            y += (int)((float) height * (1f - camera.getMaxRenderArea().getY()));
 
-                y = (int) ((height * projectedPosition.getY() * (camera.getMaxRenderArea().getY() - camera.getMinRenderArea().getY())- sprite.getScaledSprite().getHeight() / 2f));
-                y += (int)((float) height * (1f - camera.getMaxRenderArea().getY()));
 
-                frameBuffer.drawImage(sprite.getScaledSprite(), null, x, y);
-            }
+            frameBuffer.drawImage(spriteToRender, null, x, y);
         }
 
         window.repaint();
@@ -144,7 +143,6 @@ public class SoftwareRenderer {
     public void addSpriteToQueue(SpriteRenderer sprite) {
 
         sprites.add(sprite);
-        sprite.rescale();
     }
 
     /** Remove a sprite from the rendering list.
@@ -168,29 +166,12 @@ public class SoftwareRenderer {
         return aspectRatio;
     }
 
-    /** Set the camera used to render its content.
-     *
-     * @param camera The camera to use for rendering.
-     */
-    public void setCamera(Camera camera) {
-
-        this.camera = camera;
-
-        verticalSpriteSizeTarget = (float)(height) / camera.getOrthographicSize();
-
-        for (int i = 0; i < sprites.size(); i++) {
-
-            sprites.get(i).rescale();
-        }
-    }
-
     /** Get the vertical size targeted when rescaling sprites.
      *
      * @return
      */
     public float getVerticalSpriteSizeTarget() {
 
-//        return verticalSpriteSizeTarget * (camera.getMaxRenderArea().getX() - camera.getMinRenderArea().getX());
         return camera.getVerticalSpriteSizeTarget();
     }
 
@@ -201,6 +182,15 @@ public class SoftwareRenderer {
     public Camera getCamera() {
 
         return camera;
+    }
+
+    /** Set the camera used to render its content.
+     *
+     * @param camera The camera to use for rendering.
+     */
+    public void setCamera(Camera camera) {
+
+        this.camera = camera;
     }
 
     /** Get the width of the rendering zone.
